Updated OmpSs-2@OpenACC to use 64-bit indexes

+++ ompss2_openacc/timer.c
@@ -11,15 +11,15 @@
 #include <stdlib.h>
 #include <sys/time.h>
 
-uint64_t timer_ticks()
+size_t timer_ticks()
 {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
 
-	return ((uint64_t) tv.tv_sec) * 1000000 + (uint64_t) tv.tv_usec;
+	return ((size_t) tv.tv_sec) * 1000000 + (size_t) tv.tv_usec;
 }
 
-double timer_interval_seconds(uint64_t start, uint64_t end)
+double timer_interval_seconds(size_t start, size_t end)
 {
 	return (end - start) * 1.0e-6;
 }
