+++ ompss2_openacc/utilities.h
@@ -1,10 +1,9 @@
 #ifndef UTILITIES_H_
 #define UTILITIES_H_
 
-#include <string.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <math.h>
+#include <stddef.h>
+#include <stdint.h>
+#include "zpic.h"
 
 // NVIDIA GPUs
 #define DEVICE_TYPE acc_device_nvidia
@@ -22,8 +21,8 @@ void realloc_device_buffer(void **restrict ptr, const size_t old_size, const siz
 void* alloc_device_buffer(const size_t size, const int device);
 void free_device_buffer(void *ptr);
 
-void prefix_sum_openacc(int *restrict vector, const int size);
-void prefix_sum_serial(int *restrict vector, const int size);
+void prefix_sum_openacc(int64_t *restrict vector, const int64_t size);
+void prefix_sum_serial(int64_t *restrict vector, const int64_t size);
 
 #ifdef ENABLE_PREFETCH
 void grid_prefetch_openacc(t_vfld *buffer, const int size, const int device, void *stream);
